Rename types to entities
