+++ android/app/src/main/java/com/github/braillesystems/learnbraille/data/entities/StepData.kt
@@ -1,4 +1,4 @@
-package com.github.braillesystems.learnbraille.data.types
+package com.github.braillesystems.learnbraille.data.entities
 
 import androidx.room.TypeConverter
