+++ android/app/src/main/java/com/github/braillesystems/learnbraille/res/russian/Users.kt
@@ -1,6 +1,6 @@
 package com.github.braillesystems.learnbraille.res.russian
 
-import com.github.braillesystems.learnbraille.data.types.User
+import com.github.braillesystems.learnbraille.data.entities.User
 
 val PREPOPULATE_USERS = listOf(
     User(
