+++ android/app/src/main/java/com/github/braillesystems/learnbraille/res/russian/steps/Lesson2.kt
@@ -1,13 +1,13 @@
 package com.github.braillesystems.learnbraille.res.russian.steps
 
-import com.github.braillesystems.learnbraille.data.types.Info
-import com.github.braillesystems.learnbraille.data.types.InputSymbol
-import com.github.braillesystems.learnbraille.data.types.ShowSymbol
-import com.github.braillesystems.learnbraille.data.types.StepData
+import com.github.braillesystems.learnbraille.data.entities.Info
+import com.github.braillesystems.learnbraille.data.entities.InputSymbol
+import com.github.braillesystems.learnbraille.data.entities.ShowSymbol
+import com.github.braillesystems.learnbraille.data.entities.StepData
 import com.github.braillesystems.learnbraille.res.russian.symbols.symbolMap
 
 private fun Step(title: String, data: StepData) =
-    com.github.braillesystems.learnbraille.data.types.Step(
+    com.github.braillesystems.learnbraille.data.entities.Step(
         title = title,
         lessonId = 2L,
         data = data
