+++ android/app/src/main/java/com/github/braillesystems/learnbraille/res/russian/steps/LessonFirst.kt
@@ -1,7 +1,7 @@
 package com.github.braillesystems.learnbraille.res.russian.steps
 
-import com.github.braillesystems.learnbraille.data.types.FirstInfo
-import com.github.braillesystems.learnbraille.data.types.Step
+import com.github.braillesystems.learnbraille.data.entities.FirstInfo
+import com.github.braillesystems.learnbraille.data.entities.Step
 import com.github.braillesystems.learnbraille.res.russian.PREPOPULATE_LESSONS
 
 val VERY_FIRST = Step(
