+++ android/app/src/main/java/com/github/braillesystems/learnbraille/res/russian/symbols/NonLetters.kt
@@ -1,6 +1,6 @@
 package com.github.braillesystems.learnbraille.res.russian.symbols
 
-import com.github.braillesystems.learnbraille.data.types.Symbol
+import com.github.braillesystems.learnbraille.data.entities.Symbol
 
 // TODO fill punctuation
 val PUNCTUATION get() = listOf<Symbol>()
